all caches hit rate is float, not double

+++ entity-store/src/main/java/jetbrains/exodus/entitystore/EntityIterableCacheAdapter.java
@@ -69,7 +69,7 @@ final class EntityIterableCacheAdapter {
         cache.remove(key);
     }
 
-    double hitRate() {
+    float hitRate() {
         return cache.hitRate();
     }
